requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-postgres
pyspark
psycopg2-binary
requests
